rename track-bundle dir to track-reference
