+++ packages/core/src/track-reference/track-bundle.types.ts
