+++ packages/core/src/track-reference/track-bundle.utils.ts
